+++ app/middleware/body-parser.js
@@ -1,0 +1,4 @@
+module.exports = (app, conf, log) => {
+  log.debug('adding body-parser');
+  return require('body-parser').urlencoded({extended: true});
+};
