+++ app/middleware/cookie-parser.js
@@ -1,0 +1,4 @@
+module.exports = (app, conf, log) => {
+  log.debug('adding cookie-parser');
+  return require('cookie-parser')();
+};
