+++ app/middleware/raven-error-handler.js
@@ -1,0 +1,5 @@
+module.exports = (app, conf, log) => {
+  log.debug('adding raven-errorhandler');
+  const raven = require('raven');
+  return raven.errorHandler();
+};
