+++ app/middleware/raven-request-handler.js
@@ -1,0 +1,6 @@
+module.exports = (app, conf, log) => {
+  log.debug('adding raven request handler');
+  const raven = require('raven');
+  raven.config(conf.sentry.dsn, conf.sentry.options).install();
+  return raven.requestHandler();
+};
